requirements.txt
fastapi == 0.99.1
httpx
uvicorn
openai == 0.27.8
pydantic == 1.10.10
passlib
sqlite-utils
python-dotenv
